show frictionless and sensitive data alerts to curators on landing page
