+++ app/assets/stylesheets/stash_engine/frictionless-components.css
@@ -119,7 +119,7 @@
 }
 
 .frictionless-components-report .result {
-    margin: 30px 0;
+    margin: 0 0 30px;
     font-size: 12px;
     position: relative;
 }
