+++ app/controllers/stash_engine/generic_files_controller.rb
@@ -5,8 +5,9 @@ module StashEngine
   class GenericFilesController < ApplicationController
 
     before_action :setup_class_info, :require_login
-    before_action :set_file_info, only: %i[destroy_manifest rename]
+    before_action :set_file_info, only: %i[destroy_manifest frictionless_report sd_report rename]
     before_action :ajax_require_modifiable, only: %i[destroy_manifest rename validate_urls presign_upload upload_complete]
+    protect_from_forgery except: %i[frictionless_report sd_report]
 
     # apply Pundit?
 
@@ -107,7 +108,7 @@ module StashEngine
     end
 
     def trigger_sd_scan
-      scannable_files = resource.generic_files.scannable_files.where(id: params['file_ids'])
+      scannable_files = resource.data_files.scannable_files.where(id: params['file_ids'])
       render json: 'Nothing to trigger'.to_json and return if scannable_files.empty?
 
       scannable_files.each do |file|
@@ -138,6 +139,16 @@ module StashEngine
       render json: files.as_json(
         methods: %i[type uploaded], include: { frictionless_report: { only: %i[report status] } }
       )
+    end
+
+    def frictionless_report
+      @report = JSON.parse(@file.frictionless_report&.report)
+      respond_to(&:js)
+    end
+
+    def sd_report
+      @report = @file.sensitive_data_report&.report&.dig('report')
+      respond_to(&:js)
     end
 
     def rename
