+++ app/views/stash_engine/generic_files/sd_report.js.erb
@@ -1,0 +1,2 @@
+document.getElementById('genericModalContent').innerHTML = "<%= escape_javascript(render partial: 'sd_report') %>"
+document.getElementById('genericModalDialog').showModal();
