+++ app/views/stash_engine/landing/_file_admin_checks.html.erb
@@ -1,0 +1,12 @@
+<span class="input-line" style="margin-left: auto">
+<% if file.sensitive_data_report&.status == 'issues' %>
+  <%= form_with(url: "/data_files/#{file.id}/sd_report.js", method: :get, local: false, class: 'o-button__inline-form') do %>
+    <button class="o-button__plain-textlink error-text" aria-label="Sensitive data alerts" title="Sensitive data alerts"><i class="fas fa-person-circle-exclamation" aria-hidden="true"></i></button>
+  <% end %>
+<% end %>
+<% if file.frictionless_report&.status == 'issues' %>
+  <%= form_with(url: "/data_files/#{file.id}/frictionless_report.js", method: :get, local: false, class: 'o-button__inline-form') do %>
+    <button type="submit" class="o-button__plain-textlink error-text" aria-label="Tabular data check alerts" title="Tabular data check alerts"><i class="fas fa-file-circle-exclamation" aria-hidden="true"></i></button>
+  <% end %>
+<% end %>
+</span>
